% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processor_utils.R
\name{batch_process_domains}
\alias{batch_process_domains}
\title{Batch process multiple domains}
\usage{
batch_process_domains(domains, data_file, age_group = "adult")
}
\arguments{
\item{domains}{Vector of domain names}

\item{data_file}{Path to data file}

\item{age_group}{Age group}
}
\value{
A list of DomainProcessorR6 objects
}
\description{
Batch process multiple domains
}
