% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_domain_processor.R
\name{check_domain_has_data}
\alias{check_domain_has_data}
\title{Check if a domain has data in the specified file}
\usage{
check_domain_has_data(domain_names, data_file)
}
\description{
Check if a domain has data in the specified file
}
