% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConfigManagerR6.R
\name{ConfigManagerR6}
\alias{ConfigManagerR6}
\title{ConfigManagerR6 Class}
\description{
Centralized configuration management for neuro2 package
Consolidates settings from YAML files, environment variables, and defaults
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ConfigManagerR6-new}{\code{ConfigManagerR6$new()}}
\item \href{#method-ConfigManagerR6-load_default_config}{\code{ConfigManagerR6$load_default_config()}}
\item \href{#method-ConfigManagerR6-merge_configs}{\code{ConfigManagerR6$merge_configs()}}
\item \href{#method-ConfigManagerR6-apply_env_overrides}{\code{ConfigManagerR6$apply_env_overrides()}}
\item \href{#method-ConfigManagerR6-set_nested_value}{\code{ConfigManagerR6$set_nested_value()}}
\item \href{#method-ConfigManagerR6-validate_config}{\code{ConfigManagerR6$validate_config()}}
\item \href{#method-ConfigManagerR6-get}{\code{ConfigManagerR6$get()}}
\item \href{#method-ConfigManagerR6-set}{\code{ConfigManagerR6$set()}}
\item \href{#method-ConfigManagerR6-save_config}{\code{ConfigManagerR6$save_config()}}
\item \href{#method-ConfigManagerR6-clone}{\code{ConfigManagerR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$new(config_file = NULL, variables_file = "_variables.yml")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-load_default_config"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-load_default_config}{}}}
\subsection{Method \code{load_default_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$load_default_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-merge_configs"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-merge_configs}{}}}
\subsection{Method \code{merge_configs()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$merge_configs(base, override)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-apply_env_overrides"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-apply_env_overrides}{}}}
\subsection{Method \code{apply_env_overrides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$apply_env_overrides()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-set_nested_value"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-set_nested_value}{}}}
\subsection{Method \code{set_nested_value()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$set_nested_value(path, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-validate_config"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-validate_config}{}}}
\subsection{Method \code{validate_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$validate_config()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-get"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$get(path, default = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-set"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-set}{}}}
\subsection{Method \code{set()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$set(path, value)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-save_config"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-save_config}{}}}
\subsection{Method \code{save_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$save_config(file = "neuro2_config.yml")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ConfigManagerR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ConfigManagerR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ConfigManagerR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
