% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomainProcessorFactoryR6.R
\name{create_domain_processors}
\alias{create_domain_processors}
\title{Create multiple domain processors}
\usage{
create_domain_processors(
  domain_keys,
  age_group = "adult",
  include_multi_rater = TRUE,
  config = NULL
)
}
\arguments{
\item{domain_keys}{Vector of domain identifiers}

\item{age_group}{Age group ("adult" or "child")}

\item{include_multi_rater}{Whether to create multi-rater processors}

\item{config}{Optional configuration}
}
\description{
Create multiple domain processors
}
