% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_domain_processor.R
\name{create_include_list}
\alias{create_include_list}
\title{Create include list for main template}
\usage{
create_include_list(generated_files, verbose = TRUE)
}
\description{
Create include list for main template
}
