% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_temp_dir}
\alias{create_temp_dir}
\title{Create Temporary Directory with Cleanup}
\usage{
create_temp_dir(prefix = "neurotypr_", cleanup = TRUE)
}
\arguments{
\item{prefix}{Directory prefix}

\item{cleanup}{Whether to cleanup on exit}
}
\value{
Path to temporary directory
}
\description{
Creates a temporary directory that is automatically cleaned up.
}
