% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_integration.R
\name{create_workflow_template}
\alias{create_workflow_template}
\title{Create a complete workflow template}
\usage{
create_workflow_template(output_file = "neuropsych_report.qmd", verbose = TRUE)
}
\description{
Create a complete workflow template
}
