% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomainProcessorFactoryR6.R
\name{DomainProcessorFactoryR6}
\alias{DomainProcessorFactoryR6}
\title{DomainProcessorFactoryR6 Class}
\description{
Factory class for creating domain processors with smart defaults
Simplifies initialization and reduces code duplication
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DomainProcessorFactoryR6-new}{\code{DomainProcessorFactoryR6$new()}}
\item \href{#method-DomainProcessorFactoryR6-build_domain_registry}{\code{DomainProcessorFactoryR6$build_domain_registry()}}
\item \href{#method-DomainProcessorFactoryR6-create_processor}{\code{DomainProcessorFactoryR6$create_processor()}}
\item \href{#method-DomainProcessorFactoryR6-create_multi_processor}{\code{DomainProcessorFactoryR6$create_multi_processor()}}
\item \href{#method-DomainProcessorFactoryR6-get_available_raters}{\code{DomainProcessorFactoryR6$get_available_raters()}}
\item \href{#method-DomainProcessorFactoryR6-get_input_file}{\code{DomainProcessorFactoryR6$get_input_file()}}
\item \href{#method-DomainProcessorFactoryR6-batch_create}{\code{DomainProcessorFactoryR6$batch_create()}}
\item \href{#method-DomainProcessorFactoryR6-get_registry_info}{\code{DomainProcessorFactoryR6$get_registry_info()}}
\item \href{#method-DomainProcessorFactoryR6-validate_domain_data}{\code{DomainProcessorFactoryR6$validate_domain_data()}}
\item \href{#method-DomainProcessorFactoryR6-clone}{\code{DomainProcessorFactoryR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$new(config = NULL, error_handler = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-build_domain_registry"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-build_domain_registry}{}}}
\subsection{Method \code{build_domain_registry()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$build_domain_registry()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-create_processor"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-create_processor}{}}}
\subsection{Method \code{create_processor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$create_processor(
  domain_key,
  age_group = "adult",
  rater = "self",
  custom_config = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-create_multi_processor"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-create_multi_processor}{}}}
\subsection{Method \code{create_multi_processor()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$create_multi_processor(
  domain_key,
  age_group = "adult"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-get_available_raters"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-get_available_raters}{}}}
\subsection{Method \code{get_available_raters()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$get_available_raters(domain_info, age_group)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-get_input_file"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-get_input_file}{}}}
\subsection{Method \code{get_input_file()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$get_input_file(data_source)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-batch_create"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-batch_create}{}}}
\subsection{Method \code{batch_create()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$batch_create(
  domain_keys,
  age_group = "adult",
  include_multi_rater = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-get_registry_info"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-get_registry_info}{}}}
\subsection{Method \code{get_registry_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$get_registry_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-validate_domain_data"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-validate_domain_data}{}}}
\subsection{Method \code{validate_domain_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$validate_domain_data(domain_key)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorFactoryR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorFactoryR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorFactoryR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
