% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils.R
\name{.determine_input_file}
\alias{.determine_input_file}
\title{Determine input file based on phenotype}
\usage{
.determine_input_file(pheno)
}
\description{
Determine input file based on phenotype
}
\keyword{internal}
