% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils2.R
\name{.fix_emotion_child_text_files}
\alias{.fix_emotion_child_text_files}
\title{Regenerate emotion child text files with proper separation}
\usage{
.fix_emotion_child_text_files()
}
\description{
Regenerate emotion child text files with proper separation
}
