% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils.R
\name{.generate_text_results}
\alias{.generate_text_results}
\title{Generate text results for all rater types}
\usage{
.generate_text_results(data, pheno)
}
\description{
Generate text results for all rater types
}
\keyword{internal}
