% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_validation.R
\name{.get_available_domains}
\alias{.get_available_domains}
\title{Get Available Domains from Data}
\usage{
.get_available_domains(data_files, verbose = TRUE)
}
\arguments{
\item{data_files}{List of loaded data frames}

\item{verbose}{Whether to show messages}
}
\value{
Character vector of available domains
}
\description{
Extracts unique domains from loaded data files.
}
