% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils.R
\name{.get_available_raters}
\alias{.get_available_raters}
\title{Get available raters from data}
\usage{
.get_available_raters(data)
}
\arguments{
\item{data}{A data frame containing neuropsych test results}
}
\value{
Character vector of available rater types
}
\description{
This function looks at your data and figures out which types of
raters (self, parent, teacher, observer) actually have data. It's like taking
attendance - who actually showed up with data?
}
\keyword{internal}
