% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_type_utils.R
\name{.get_score_groups}
\alias{.get_score_groups}
\title{Get score type groups for test names}
\usage{
.get_score_groups(test_names)
}
\arguments{
\item{test_names}{Character vector of test names}
}
\value{
Named list of score type groups
}
\description{
Get score type groups for test names
}
