% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.neuro2_config}
\alias{.neuro2_config}
\title{Allow users to configure output directories}
\usage{
.neuro2_config(
  figs_dir = "figs",
  output_dir = "output",
  tmp_dir = "tmp",
  base_dir = "."
)
}
\description{
Allow users to configure output directories
}
