% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrilldownR6.R
\name{drilldown}
\alias{drilldown}
\title{Drilldown on Neuropsych Domains (Function Wrapper)}
\usage{
drilldown(
  data,
  patient,
  neuro_domain = c("Neuropsychological Test Scores", "Behavioral Rating Scales",
    "Effort/Validity Test Scores"),
  theme
)
}
\arguments{
\item{data}{Dataset to use.}

\item{patient}{Name of patient.}

\item{neuro_domain}{Name of neuropsych domain to add to HC series.}

\item{theme}{The highcharter theme to use.}
}
\value{
A drilldown plot
}
\description{
This function uses the R Highcharter package and drilldown function to
"drilldown" on neuropsychological domains and test scores. It's a wrapper
around the DrilldownR6 class.
}
