% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{drilldown2}
\alias{drilldown2}
\title{Drilldown on PASS Verbal/Nonverbal Domains
This function creates a drilldown visualization for PASS Verbal/Nonverbal components.}
\usage{
drilldown2(
  data,
  patient,
  neuro_domain = c("Neuropsychological Test Scores", "Behavioral Rating Scales",
    "Effort/Validity Test Scores"),
  theme
)
}
\arguments{
\item{data}{Dataset to use.}

\item{patient}{Name of patient.}

\item{neuro_domain}{Name of neuropsych domain to add to HC series.}

\item{theme}{The highcharter theme to use.}
}
\value{
A drilldown plot
}
\description{
Drilldown on PASS Verbal/Nonverbal Domains
This function creates a drilldown visualization for PASS Verbal/Nonverbal components.
}
