% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrilldownR6.R
\name{DrilldownR6}
\alias{DrilldownR6}
\title{DrilldownR6 Class}
\description{
DrilldownR6 Class

DrilldownR6 Class
}
\details{
An R6 class that creates interactive
highcharter drilldown plots for neuropsychological domains.
This is an R6 implementation of the drilldown function
with identical functionality.
}
\section{Methods}{

\describe{
\item{initialize}{Initialize a new DrilldownR6 object
with configuration and data.}
\item{create_plot}{Generate the interactive highcharter
drilldown plot based on the object's configuration.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{Dataset to use.}

\item{\code{patient}}{Name of patient.}

\item{\code{neuro_domain}}{Name of neuropsych domain to add to HC series.}

\item{\code{theme}}{The highcharter theme to use.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DrilldownR6-new}{\code{DrilldownR6$new()}}
\item \href{#method-DrilldownR6-create_plot}{\code{DrilldownR6$create_plot()}}
\item \href{#method-DrilldownR6-clone}{\code{DrilldownR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrilldownR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-DrilldownR6-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DrilldownR6 object with configuration and data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrilldownR6$new(
  data,
  patient,
  neuro_domain = c("Neuropsychological Test Scores", "Behavioral Rating Scales",
    "Effort/Validity Test Scores"),
  theme
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{Dataset to use.}

\item{\code{patient}}{Name of patient.}

\item{\code{neuro_domain}}{Name of neuropsych domain to add to HC series.}

\item{\code{theme}}{The highcharter theme to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new DrilldownR6 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrilldownR6-create_plot"></a>}}
\if{latex}{\out{\hypertarget{method-DrilldownR6-create_plot}{}}}
\subsection{Method \code{create_plot()}}{
Generate the interactive highcharter
drilldown plot based on the object's configuration.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrilldownR6$create_plot()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A highcharter object representing the interactive drilldown plot.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DrilldownR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DrilldownR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DrilldownR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
