% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_type_cache.R
\name{ensure_score_type_utils}
\alias{ensure_score_type_utils}
\title{Ensure Score Type Utils}
\usage{
ensure_score_type_utils()
}
\description{
Ensures that the score type utilities are loaded and available
}
