% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{filter_data}
\alias{filter_data}
\title{Utility functions to replace NeurotypR dependencies}
\usage{
filter_data(data, domain = NULL, scale = NULL)
}
\arguments{
\item{data}{Data frame to filter}

\item{domain}{Domain(s) to filter by}

\item{scale}{Scale(s) to filter by}
}
\value{
Filtered data frame
}
\description{
This file contains utility functions that replace NeurotypR functions
used in the neuro2 package.
Filter data by domain and scale
}
