% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{filter_data2}
\alias{filter_data2}
\title{Filters Data by Domain and Scale}
\usage{
filter_data2(data, domains = NULL, scales = NULL)
}
\arguments{
\item{data}{A dataframe or tibble}

\item{domains}{Character vector of domain names to filter by}

\item{scales}{Character vector of scale names to filter by}
}
\value{
Returns a filtered data frame
}
\description{
This function filters a dataframe by domain and scale.
}
