% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils2.R
\name{generate_emotion_rater_outputs}
\alias{generate_emotion_rater_outputs}
\title{Generate outputs for emotion domain with multiple raters}
\usage{
generate_emotion_rater_outputs(data, processor)
}
\description{
Generate outputs for emotion domain with multiple raters
}
