% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils2.R
\name{generate_standard_outputs}
\alias{generate_standard_outputs}
\title{Generate standard outputs for single-rater domains}
\usage{
generate_standard_outputs(data, pheno, rater = NULL)
}
\description{
Generate standard outputs for single-rater domains
}
