% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_domain_generator.R
\name{generate_workflow_domains}
\alias{generate_workflow_domains}
\title{Generate Workflow Domains}
\usage{
generate_workflow_domains(config)
}
\arguments{
\item{config}{Configuration list from load_workflow_config}
}
\value{
Logical indicating success
}
\description{
Generate Workflow Domains
}
