% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_report_generator.R
\name{generate_workflow_report}
\alias{generate_workflow_report}
\title{Generate Workflow Report}
\usage{
generate_workflow_report(config)
}
\arguments{
\item{config}{Configuration list from load_workflow_config}
}
\value{
Logical indicating success
}
\description{
Generate Workflow Report
}
