% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_type_utils.R
\name{get_all_score_type_notes}
\alias{get_all_score_type_notes}
\title{Get All Score Type Notes}
\usage{
get_all_score_type_notes()
}
\value{
A list containing all score type notes
}
\description{
Returns a list of all available score type notes
}
