% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processor_utils.R
\name{get_domain_info}
\alias{get_domain_info}
\title{Get domain information from the lookup table}
\usage{
get_domain_info(test_lookup_file = "test_testname_rater.csv")
}
\arguments{
\item{test_lookup_file}{Path to test lookup file}
}
\value{
A data frame with domain summary information
}
\description{
Get domain information from the lookup table
}
