% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_validation_utils.R
\name{get_domains_with_data}
\alias{get_domains_with_data}
\title{Get Domains With Data}
\usage{
get_domains_with_data(neurocog_data, neurobehav_data, domain_config)
}
\arguments{
\item{neurocog_data}{Neurocognitive data}

\item{neurobehav_data}{Neurobehavioral data}

\item{domain_config}{Domain configuration list}
}
\value{
List of domains with data
}
\description{
Returns only domains that have actual data
}
