% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorHandlerR6_fixed.R
\name{get_error_handler}
\alias{get_error_handler}
\title{Get or create global error handler}
\usage{
get_error_handler(config = NULL)
}
\arguments{
\item{config}{Optional configuration object}
}
\description{
Get or create global error handler
}
