% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_neuropsych_loader.R
\name{get_example_queries}
\alias{get_example_queries}
\title{Example DuckDB queries for neuropsychological data}
\usage{
get_example_queries()
}
\value{
List of example queries
}
\description{
Example DuckDB queries for neuropsychological data
}
