% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_type_utils.R
\name{get_score_footnotes}
\alias{get_score_footnotes}
\title{Get footnotes for score types}
\usage{
get_score_footnotes(score_types)
}
\arguments{
\item{score_types}{Character vector of score types}
}
\value{
Named list of footnotes
}
\description{
Get footnotes for score types
}
