% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_type_utils.R
\name{get_source_note_by_score_type}
\alias{get_source_note_by_score_type}
\title{Get Source Note Based on Score Type}
\usage{
get_source_note_by_score_type(data, default_note = NULL)
}
\arguments{
\item{data}{A data frame containing a score_type column}

\item{default_note}{The default source note to use if no score types are found}
}
\value{
A character string containing the appropriate source note
}
\description{
Determines the appropriate source note based on the score types present in the data
}
