% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf.R
\name{gpluck_compute_percentile_range}
\alias{gpluck_compute_percentile_range}
\title{Compute Percentile & Performance Range}
\usage{
gpluck_compute_percentile_range(
  .data,
  score_col,
  score_type = c("z_score", "scaled_score", "t_score", "standard_score"),
  keep_intermediate = FALSE
)
}
\arguments{
\item{.data}{A data.frame/tibble.}

\item{score_col}{Unquoted name of the score column in \code{.data}.}

\item{score_type}{One of "z_score", "scaled_score", "t_score", "standard_score".}

\item{keep_intermediate}{Keep pct1/pct2/pct3 columns? Default FALSE.}
}
\value{
\code{.data} with added columns: z, percentile, range.
}
\description{
Compute Percentile & Performance Range
}
