% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_neuro2.R
\name{init_neuro2_workspace}
\alias{init_neuro2_workspace}
\title{Initialize neuro2 workspace}
\usage{
init_neuro2_workspace(patient_name = NULL, verbose = TRUE)
}
\description{
Sets up directories and loads functionality
}
