% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{iq}
\alias{iq}
\title{Internal neuropsychological test data - IQ}
\format{
A data frame with neuropsychological test results
}
\description{
Internal dataset containing IQ test results data used for report generation.
This data is not exported and is for internal package use only.
}
\keyword{datasets}
\keyword{internal}
