% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_upload_helpers.R
\name{list_pdf_templates}
\alias{list_pdf_templates}
\title{List available PDF extraction templates}
\usage{
list_pdf_templates()
}
\value{
Character vector of available test types for PDF extraction
}
\description{
List available PDF extraction templates
}
