% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_batch_openai_merged.R, R/neuro2_llm_bot.R
\name{llm_cache_dir}
\alias{llm_cache_dir}
\title{LLM Cache Directory}
\usage{
llm_cache_dir()

llm_cache_dir()
}
\value{
Character string of the cache directory path.

Character string of the cache directory path.
}
\description{
Returns the path to the LLM cache directory, creating it if it doesn't exist.

Returns the path to the LLM cache directory, creating it if it doesn't exist.
}
