% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_components_once.R
\name{load_components_once}
\alias{load_components_once}
\title{Centralized Component Loader - Runs ONCE}
\description{
This replaces all the source() calls scattered in functions
Load this ONCE at the start of your workflow
}
