% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_neuro2.R
\name{load_neuro2}
\alias{load_neuro2}
\title{Load neuro2 package functionality}
\usage{
load_neuro2(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to show loading messages}
}
\description{
Load neuro2 package functionality
}
