% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages.R
\name{load_neuropsych_packages}
\alias{load_neuropsych_packages}
\title{Load all required packages for neuropsych reports}
\usage{
load_neuropsych_packages(verbose = TRUE)
}
\arguments{
\item{verbose}{Whether to print loading messages}
}
\value{
Invisible NULL
}
\description{
This function loads the minimal set of packages needed for
report generation. Most data processing is handled by R6 classes.
}
