% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_utils.R
\name{log_message}
\alias{log_message}
\title{Log Message}
\usage{
log_message(message, type = "INFO", log_file = "workflow.log")
}
\arguments{
\item{message}{Message to log}

\item{type}{Type of message (INFO, WARNING, ERROR, etc)}

\item{log_file}{Path to log file}
}
\description{
Log Message
}
