% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_test_data.R
\name{merge_subtests}
\alias{merge_subtests}
\title{Function to merge multiple CSV files from the same test}
\usage{
merge_subtests(test, suffix = "csv", directory = here::here("data", "csv"))
}
\arguments{
\item{test}{Character string with the test prefix for files to merge}

\item{suffix}{Character string with the file suffix (default: "csv")}

\item{directory}{Character string with the directory containing the files}
}
\value{
A data frame containing the merged data
}
\description{
Function to merge multiple CSV files from the same test
}
