% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro2-package.R
\docType{package}
\name{neuro2-package}
\alias{neuro2}
\alias{neuro2-package}
\title{neuro2: Quarto + Typst Report Generator}
\description{
Provides Quarto .qmd and Typst .typ templates for neuropsychological evaluation data processing and report generation.
}
\author{
\strong{Maintainer}: Joey Trampush \email{joey.trampush@brainworkup.org}

}
\keyword{internal}
