% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neuro2MainR6.R
\name{neuro2_quick_start}
\alias{neuro2_quick_start}
\title{Quick start function for new users}
\usage{
neuro2_quick_start(patient_name, age, data_dir = "data-raw/csv")
}
\arguments{
\item{patient_name}{Patient name}

\item{age}{Patient age}

\item{data_dir}{Directory containing CSV data files}
}
\description{
Quick start function for new users
}
