% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neuro2MainR6.R
\name{neuro2_workflow}
\alias{neuro2_workflow}
\title{Create and run neuropsychological report workflow}
\usage{
neuro2_workflow(
  config_file = "config.yml",
  variables_file = "_variables.yml",
  domains = NULL,
  age_group = NULL,
  load_data = TRUE,
  generate_report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{config_file}{Path to configuration file}

\item{variables_file}{Path to variables file}

\item{domains}{Vector of domains to process}

\item{age_group}{Age group ("adult" or "child")}

\item{load_data}{Whether to load data}

\item{generate_report}{Whether to generate report}

\item{verbose}{Whether to show progress messages}
}
\description{
Create and run neuropsychological report workflow
}
