% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{neurobehav}
\alias{neurobehav}
\title{Internal neuropsychological test data - Neurobehavioral}
\format{
A data frame with neurobehavioral assessment results
}
\description{
Internal dataset containing neurobehavioral assessment data used for report generation.
This data is not exported and is for internal package use only.
}
\keyword{datasets}
\keyword{internal}
