% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{neurotypr_utils}
\alias{neurotypr_utils}
\alias{\%||\%}
\title{Enhanced Utility Functions for NeurotypR}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Primary value}

\item{y}{Default value if x is NULL}
}
\description{
Core utility functions that provide common operations used throughout
the NeurotypR package. These functions improve performance, reduce
code duplication, and provide consistent behavior.

Returns the first non-null value from left to right.
Similar to ?? operator in other languages.
}
\examples{
# Returns "default" because x is NULL
x <- NULL
x \%||\% "default"

# Returns "value" because x is not NULL
x <- "value"
x \%||\% "default"
}
\keyword{Coalescing}
\keyword{Null}
\keyword{Operator}
\keyword{internal}
