% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{null-coalescing-operator}
\alias{null-coalescing-operator}
\alias{\%||\%}
\title{Null Coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{Primary value}

\item{y}{Default value if x is NULL}
}
\description{
Returns the first non-null value from left to right.
Similar to ?? operator in other languages.
}
\examples{
# Returns "default" because x is NULL
x <- NULL
x \%||\% "default"

# Returns "value" because x is not NULL
x <- "value"
x \%||\% "default"
}
