% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuropsych_test_scoring.R
\name{pegboard_nondominant_hand}
\alias{pegboard_nondominant_hand}
\title{Grooved Pegboard, NonDominant Hand}
\usage{
pegboard_nondominant_hand(age, raw_score)
}
\arguments{
\item{age}{A numeric value of the age of the participant. Must be between 16 and 89 years old.}

\item{raw_score}{A numeric value of the raw score from the GPT nondom.}
}
\value{
Returns a numeric value of the T-score for the GPT nondom hand trial.
}
\description{
Calculates a GPT dominant hand (T-score) from a raw score and age. The predicted score and standard deviation are also calculated.
}
