% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_domain_processor.R
\name{process_all_domains}
\alias{process_all_domains}
\title{Process all domains with available data}
\usage{
process_all_domains(config, patient_type, data_status)
}
\description{
Process all domains with available data
}
