% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processor_utils.R
\name{process_simple_domain}
\alias{process_simple_domain}
\title{Process a simple domain (single rater)}
\usage{
process_simple_domain(domain_name, data_file, age_group = "adult")
}
\arguments{
\item{domain_name}{Domain name}

\item{data_file}{Path to data file}

\item{age_group}{Age group ("adult" or "child")}
}
\value{
A DomainProcessorR6 object
}
\description{
Process a simple domain (single rater)
}
