% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ErrorHandlerR6.R
\name{safe_execute}
\alias{safe_execute}
\title{Safe execution wrapper function}
\usage{
safe_execute(expr, context = "operation", fallback = NULL, silent = FALSE)
}
\arguments{
\item{expr}{Expression to execute safely}

\item{context}{Context description for errors}

\item{fallback}{Value to return on error}

\item{silent}{Whether to suppress error messages}
}
\description{
Safe execution wrapper function
}
