% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_sysdata_update_fixed.R
\name{safe_use_data_internal}
\alias{safe_use_data_internal}
\title{Safely update internal package data without overwriting existing variables}
\usage{
safe_use_data_internal(
  ...,
  file = "R/sysdata.rda",
  overwrite = NULL,
  add_only = FALSE
)
}
\arguments{
\item{...}{Named objects to save}

\item{file}{Path to the .rda file (default: "R/sysdata.rda")}

\item{overwrite}{Character vector of object names that should be overwritten}

\item{add_only}{Logical, if TRUE only add new objects, never overwrite}
}
\description{
Safely update internal package data without overwriting existing variables
}
