% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_type_cache.R
\name{ScoreTypeCacheR6}
\alias{ScoreTypeCacheR6}
\title{ScoreTypeCacheR6 - Cached Score Type Manager}
\description{
Manages score type mappings with caching to avoid repetitive lookups
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ScoreTypeCacheR6-new}{\code{ScoreTypeCacheR6$new()}}
\item \href{#method-ScoreTypeCacheR6-build_mappings}{\code{ScoreTypeCacheR6$build_mappings()}}
\item \href{#method-ScoreTypeCacheR6-get_score_groups}{\code{ScoreTypeCacheR6$get_score_groups()}}
\item \href{#method-ScoreTypeCacheR6-get_footnotes}{\code{ScoreTypeCacheR6$get_footnotes()}}
\item \href{#method-ScoreTypeCacheR6-is_multi_score_battery}{\code{ScoreTypeCacheR6$is_multi_score_battery()}}
\item \href{#method-ScoreTypeCacheR6-clone}{\code{ScoreTypeCacheR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTypeCacheR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTypeCacheR6-new}{}}}
\subsection{Method \code{new()}}{
Initialize the cache
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreTypeCacheR6$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTypeCacheR6-build_mappings"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTypeCacheR6-build_mappings}{}}}
\subsection{Method \code{build_mappings()}}{
Build the score type mappings (runs once)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreTypeCacheR6$build_mappings()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTypeCacheR6-get_score_groups"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTypeCacheR6-get_score_groups}{}}}
\subsection{Method \code{get_score_groups()}}{
Get score type groups for specific test names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreTypeCacheR6$get_score_groups(test_names)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_names}}{Vector of test names to classify}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of score type groups
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTypeCacheR6-get_footnotes"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTypeCacheR6-get_footnotes}{}}}
\subsection{Method \code{get_footnotes()}}{
Get footnotes for score types
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreTypeCacheR6$get_footnotes(score_types)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{score_types}}{Vector of score types needed}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of footnotes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTypeCacheR6-is_multi_score_battery"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTypeCacheR6-is_multi_score_battery}{}}}
\subsection{Method \code{is_multi_score_battery()}}{
Check if a test uses multiple score types (like RBANS, WISC-V)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreTypeCacheR6$is_multi_score_battery(test_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{test_name}}{Test name to check}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Boolean
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ScoreTypeCacheR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ScoreTypeCacheR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ScoreTypeCacheR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
