% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflow_utils.R
\name{setup_logging}
\alias{setup_logging}
\title{Setup Logging}
\usage{
setup_logging(log_file = "workflow.log")
}
\arguments{
\item{log_file}{Path to log file}
}
\value{
Log file path
}
\description{
Setup Logging
}
