% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PackageManagerR6_fixed.R
\name{setup_neuro2_packages}
\alias{setup_neuro2_packages}
\title{Set up neuro2 package environment}
\usage{
setup_neuro2_packages(
  install_missing = FALSE,
  include_optional = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{install_missing}{Whether to install missing packages}

\item{include_optional}{Whether to check optional packages}

\item{verbose}{Whether to show progress}
}
\description{
Set up neuro2 package environment
}
