% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tbl_gt}
\alias{tbl_gt}
\title{Create a GT table}
\usage{
tbl_gt(
  data,
  pheno,
  table_name = NULL,
  vertical_padding = 0,
  source_note = NULL,
  dynamic_grp = NULL,
  multiline = TRUE
)
}
\arguments{
\item{data}{Data frame to display in the table}

\item{pheno}{Phenotype name}

\item{table_name}{Table name}

\item{vertical_padding}{Vertical padding}

\item{source_note}{Source note}

\item{dynamic_grp}{Dynamic group}

\item{multiline}{Whether to allow multiline text}
}
\value{
A GT table object
}
\description{
Create a GT table
}
