% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TableGT.R
\name{tbl_gt2}
\alias{tbl_gt2}
\title{Make Table Using gt Package for Neurocognitive Domains (Function Wrapper)}
\usage{
tbl_gt2(
  data,
  pheno = NULL,
  table_name = NULL,
  source_note = NULL,
  names = NULL,
  title = NULL,
  tab_stubhead = NULL,
  caption = NULL,
  process_md = FALSE,
  fn_scaled_score = NULL,
  fn_standard_score = NULL,
  fn_t_score = NULL,
  fn_z_score = NULL,
  fn_raw_score = NULL,
  grp_scaled_score = NULL,
  grp_standard_score = NULL,
  grp_t_score = NULL,
  grp_z_score = NULL,
  grp_raw_score = NULL,
  dynamic_grp = NULL,
  vertical_padding = NULL,
  multiline = TRUE,
  ...
)
}
\arguments{
\item{data}{A data frame to use in table generation.}

\item{pheno}{A string specifying the phenotype identifier (default NULL).}

\item{table_name}{File-safe name for saving the table (default NULL).}

\item{source_note}{A footnote to be added as a source note (default NULL).}

\item{names}{Optional renamed column headers (default NULL).}

\item{title}{Table title (default NULL).}

\item{tab_stubhead}{Label for the stubhead (default NULL).}

\item{caption}{Optional caption text (default NULL).}

\item{process_md}{Logical. Whether to parse markdown syntax in labels or values (default FALSE).}

\item{fn_scaled_score}{Footnote for scaled score (default NULL).}

\item{fn_standard_score}{Footnote for standard score (default NULL).}

\item{fn_t_score}{Footnote for t score (default NULL).}

\item{fn_z_score}{Footnote for z score (default NULL).}

\item{fn_raw_score}{Footnote for raw scores (default NULL).}

\item{grp_scaled_score}{Groups for scaled score (default NULL).}

\item{grp_standard_score}{Groups for standard score (default NULL).}

\item{grp_t_score}{Groups for t score (default NULL).}

\item{grp_z_score}{Groups for z score (default NULL).}

\item{grp_raw_score}{Groups for raw scores (default NULL).}

\item{dynamic_grp}{Named list of valid score types per group (default NULL).}

\item{vertical_padding}{Numeric scale (0–3) to control padding (default NULL).}

\item{multiline}{Logical; whether to use multiline footnotes (default TRUE).}

\item{...}{Additional arguments to be passed to the function.}
}
\value{
A formatted table with domain counts.
}
\description{
A function wrapper around the TableGT2 R6 class for backward compatibility.
}
