% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{time_it}
\alias{time_it}
\title{Time Function Execution}
\usage{
time_it(expr, message = NULL)
}
\arguments{
\item{expr}{Expression to time}

\item{message}{Optional message to display}
}
\value{
Result of expression
}
\description{
Times the execution of a function and optionally prints the duration.
}
