% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowRunnerR6.R
\name{WorkflowRunnerR6}
\alias{WorkflowRunnerR6}
\title{WorkflowRunnerR6}
\description{
R6 class for running neuro2 workflows
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{Configuration list for the workflow}

\item{\code{patient_name}}{Name of the patient for the workflow}

\item{\code{log_file}}{Path to the log file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkflowRunnerR6-new}{\code{WorkflowRunnerR6$new()}}
\item \href{#method-WorkflowRunnerR6-setup_environment}{\code{WorkflowRunnerR6$setup_environment()}}
\item \href{#method-WorkflowRunnerR6-process_data}{\code{WorkflowRunnerR6$process_data()}}
\item \href{#method-WorkflowRunnerR6-generate_domains}{\code{WorkflowRunnerR6$generate_domains()}}
\item \href{#method-WorkflowRunnerR6-generate_report}{\code{WorkflowRunnerR6$generate_report()}}
\item \href{#method-WorkflowRunnerR6-run}{\code{WorkflowRunnerR6$run()}}
\item \href{#method-WorkflowRunnerR6-print_summary}{\code{WorkflowRunnerR6$print_summary()}}
\item \href{#method-WorkflowRunnerR6-clone}{\code{WorkflowRunnerR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-new}{}}}
\subsection{Method \code{new()}}{
Initialize the workflow runner
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$new(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Configuration list}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-setup_environment"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-setup_environment}{}}}
\subsection{Method \code{setup_environment()}}{
Set up the environment for workflow processing
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$setup_environment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-process_data"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-process_data}{}}}
\subsection{Method \code{process_data()}}{
Process data for the workflow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$process_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-generate_domains"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-generate_domains}{}}}
\subsection{Method \code{generate_domains()}}{
Generate domain files
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$generate_domains()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-generate_report"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-generate_report}{}}}
\subsection{Method \code{generate_report()}}{
Generate final report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$generate_report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-run"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-run}{}}}
\subsection{Method \code{run()}}{
Run the complete workflow
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$run()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-print_summary"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-print_summary}{}}}
\subsection{Method \code{print_summary()}}{
Print workflow summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$print_summary(success)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{success}}{Whether the workflow succeeded}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
