% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duckdb_neuropsych_loader.R, R/tidy_data.R
\name{calculate_z_stats}
\alias{calculate_z_stats}
\title{Helper function to calculate z-statistics}
\usage{
calculate_z_stats(data, group_vars)

calculate_z_stats(data, group_vars)
}
\arguments{
\item{data}{A dataframe containing z-scores}

\item{group_vars}{Character vector of column names to group by}

\item{groups}{Character vector of grouping variables}
}
\value{
Data frame with added z-statistics

Dataframe with added z-score statistics columns
}
\description{
Efficiently calculates mean and standard deviation of z-scores for specified grouping variables
}
\keyword{internal}
