% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processor_utils_fixed.R
\name{check_domain_raters}
\alias{check_domain_raters}
\title{Check available raters for a domain and age group}
\usage{
check_domain_raters(domain_name, age_group = "adult", test_lookup_file = NULL)
}
\arguments{
\item{domain_name}{Domain name}

\item{age_group}{Age group}

\item{test_lookup_file}{Path to test lookup file (optional)}
}
\value{
A data frame with rater information
}
\description{
Check available raters for a domain and age group
}
