% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DomainProcessorR6Combo.R
\name{DomainProcessorR6Combo}
\alias{DomainProcessorR6Combo}
\title{DomainProcessorR6Combo - Following Memory Template Exactly}
\description{
DomainProcessorR6Combo - Following Memory Template Exactly

DomainProcessorR6Combo - Following Memory Template Exactly
}
\details{
A working implementation that generates QMD files following the exact
structure of the memory template file.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{domains}}{Character vector of domain names to process.}

\item{\code{pheno}}{Target phenotype identifier string.}

\item{\code{input_file}}{Path to the input data file.}

\item{\code{output_dir}}{Directory where output files will be saved.}

\item{\code{number}}{Domain number for file naming.}

\item{\code{data}}{The loaded and processed data.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DomainProcessorR6Combo-new}{\code{DomainProcessorR6Combo$new()}}
\item \href{#method-DomainProcessorR6Combo-load_data}{\code{DomainProcessorR6Combo$load_data()}}
\item \href{#method-DomainProcessorR6Combo-filter_by_domain}{\code{DomainProcessorR6Combo$filter_by_domain()}}
\item \href{#method-DomainProcessorR6Combo-select_columns}{\code{DomainProcessorR6Combo$select_columns()}}
\item \href{#method-DomainProcessorR6Combo-save_data}{\code{DomainProcessorR6Combo$save_data()}}
\item \href{#method-DomainProcessorR6Combo-has_multiple_raters}{\code{DomainProcessorR6Combo$has_multiple_raters()}}
\item \href{#method-DomainProcessorR6Combo-detect_emotion_type}{\code{DomainProcessorR6Combo$detect_emotion_type()}}
\item \href{#method-DomainProcessorR6Combo-generate_domain_qmd}{\code{DomainProcessorR6Combo$generate_domain_qmd()}}
\item \href{#method-DomainProcessorR6Combo-generate_standard_qmd}{\code{DomainProcessorR6Combo$generate_standard_qmd()}}
\item \href{#method-DomainProcessorR6Combo-generate_adhd_adult_qmd}{\code{DomainProcessorR6Combo$generate_adhd_adult_qmd()}}
\item \href{#method-DomainProcessorR6Combo-generate_adhd_child_qmd}{\code{DomainProcessorR6Combo$generate_adhd_child_qmd()}}
\item \href{#method-DomainProcessorR6Combo-generate_emotion_child_qmd}{\code{DomainProcessorR6Combo$generate_emotion_child_qmd()}}
\item \href{#method-DomainProcessorR6Combo-generate_emotion_adult_qmd}{\code{DomainProcessorR6Combo$generate_emotion_adult_qmd()}}
\item \href{#method-DomainProcessorR6Combo-process}{\code{DomainProcessorR6Combo$process()}}
\item \href{#method-DomainProcessorR6Combo-clone}{\code{DomainProcessorR6Combo$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-new"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new DomainProcessorR6Combo object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$new(
  domains,
  pheno,
  input_file,
  output_dir = "data",
  number = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-load_data"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-load_data}{}}}
\subsection{Method \code{load_data()}}{
Load data from the specified input file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$load_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-filter_by_domain"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-filter_by_domain}{}}}
\subsection{Method \code{filter_by_domain()}}{
Filter data to include only the specified domains
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$filter_by_domain()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-select_columns"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-select_columns}{}}}
\subsection{Method \code{select_columns()}}{
Select relevant columns from the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$select_columns()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-save_data"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-save_data}{}}}
\subsection{Method \code{save_data()}}{
Save the processed data to a file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$save_data(filename = NULL, format = "parquet")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-has_multiple_raters"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-has_multiple_raters}{}}}
\subsection{Method \code{has_multiple_raters()}}{
Check if domain has multiple raters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$has_multiple_raters()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-detect_emotion_type"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-detect_emotion_type}{}}}
\subsection{Method \code{detect_emotion_type()}}{
Detect emotion type (child/adult)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$detect_emotion_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-generate_domain_qmd"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-generate_domain_qmd}{}}}
\subsection{Method \code{generate_domain_qmd()}}{
Generate domain QMD file following the memory template exactly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$generate_domain_qmd(
  domain_name = NULL,
  output_file = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-generate_standard_qmd"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-generate_standard_qmd}{}}}
\subsection{Method \code{generate_standard_qmd()}}{
Generate standard domain QMD following the memory template structure exactly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$generate_standard_qmd(domain_name, output_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-generate_adhd_adult_qmd"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-generate_adhd_adult_qmd}{}}}
\subsection{Method \code{generate_adhd_adult_qmd()}}{
Generate ADHD adult QMD file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$generate_adhd_adult_qmd(domain_name, output_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-generate_adhd_child_qmd"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-generate_adhd_child_qmd}{}}}
\subsection{Method \code{generate_adhd_child_qmd()}}{
Generate ADHD child QMD file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$generate_adhd_child_qmd(domain_name, output_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-generate_emotion_child_qmd"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-generate_emotion_child_qmd}{}}}
\subsection{Method \code{generate_emotion_child_qmd()}}{
Generate emotion child QMD file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$generate_emotion_child_qmd(domain_name, output_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-generate_emotion_adult_qmd"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-generate_emotion_adult_qmd}{}}}
\subsection{Method \code{generate_emotion_adult_qmd()}}{
Generate emotion adult QMD file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$generate_emotion_adult_qmd(domain_name, output_file)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-process"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-process}{}}}
\subsection{Method \code{process()}}{
Run the complete processing pipeline
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$process(generate_domain_files = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DomainProcessorR6Combo-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DomainProcessorR6Combo-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DomainProcessorR6Combo$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
