% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DotplotR6.R
\name{dotplot2}
\alias{dotplot2}
\title{dotplot (convenience wrapper)}
\usage{
dotplot2(
  data,
  x,
  y,
  linewidth = 0.5,
  fill = x,
  shape = 21,
  point_size = 6,
  line_color = "black",
  colors = NULL,
  theme = "fivethirtyeight",
  return_plot = TRUE,
  filename = NULL,
  width = 10,
  height = NULL,
  base_height = 0.4,
  min_height = 4,
  height_per_row = 0.7,
  height_padding = 0.8,
  plot_title = NULL,
  domain = NULL,
  subdomain = NULL,
  narrow = NULL,
  ...
)
}
\arguments{
\item{data, x, y}{See DotplotR6 fields.}

\item{...}{Ignored.}
}
\value{
A ggplot object (invisibly NULL if \code{return_plot = FALSE}).
}
\description{
A functional wrapper around DotplotR6 for quick plots.
}
