% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExecutionTrackerR6.R
\name{ExecutionTrackerR6}
\alias{ExecutionTrackerR6}
\title{Execution Tracker - Prevents Multiple Runs}
\description{
Use this to track what has been executed and prevent re-runs
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ExecutionTrackerR6-new}{\code{ExecutionTrackerR6$new()}}
\item \href{#method-ExecutionTrackerR6-can_execute}{\code{ExecutionTrackerR6$can_execute()}}
\item \href{#method-ExecutionTrackerR6-mark_executed}{\code{ExecutionTrackerR6$mark_executed()}}
\item \href{#method-ExecutionTrackerR6-reset}{\code{ExecutionTrackerR6$reset()}}
\item \href{#method-ExecutionTrackerR6-get_summary}{\code{ExecutionTrackerR6$get_summary()}}
\item \href{#method-ExecutionTrackerR6-clone}{\code{ExecutionTrackerR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionTrackerR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionTrackerR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionTrackerR6$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionTrackerR6-can_execute"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionTrackerR6-can_execute}{}}}
\subsection{Method \code{can_execute()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionTrackerR6$can_execute(task_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionTrackerR6-mark_executed"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionTrackerR6-mark_executed}{}}}
\subsection{Method \code{mark_executed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionTrackerR6$mark_executed(task_id, status = "success")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionTrackerR6-reset"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionTrackerR6-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionTrackerR6$reset()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionTrackerR6-get_summary"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionTrackerR6-get_summary}{}}}
\subsection{Method \code{get_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionTrackerR6$get_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ExecutionTrackerR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ExecutionTrackerR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ExecutionTrackerR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
