% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R, R/utils.R
\name{filter_data}
\alias{filter_data}
\title{Filters Data by Domain and Scale}
\usage{
filter_data(data, domain = NULL, scale = NULL)

filter_data(data, domain = NULL, scale = NULL)
}
\arguments{
\item{data}{Data frame to filter}

\item{domain}{Domain(s) to filter by}

\item{scale}{Scale(s) to filter by}

\item{domains}{Character vector of domain names to filter by}

\item{scales}{Character vector of scale names to filter by}
}
\value{
Returns a filtered data frame

Filtered data frame
}
\description{
This function filters a dataframe by domain and scale.

This file contains utility functions that replace NeurotypR functions
used in the neuro2 package.
Filter data by domain and scale
}
