% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processor_utils_fixed.R
\name{get_domain_info}
\alias{get_domain_info}
\title{Get domain information from test lookup or registry}
\usage{
get_domain_info(test_lookup_file = NULL, use_factory = TRUE)
}
\arguments{
\item{test_lookup_file}{Path to test lookup file (optional)}

\item{use_factory}{Whether to use factory registry}
}
\value{
A data frame with domain summary information
}
\description{
Get domain information from test lookup or registry
}
