% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro2_llm.R
\name{get_domains_with_data}
\alias{get_domains_with_data}
\title{Get Domains With Data}
\usage{
get_domains_with_data()
}
\value{
Character vector of domain names that have data
}
\description{
Determines which domains actually have data available
}
