% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_neuro2.R
\name{init_neuro2_workspace}
\alias{init_neuro2_workspace}
\title{Initialize neuro2 workspace}
\usage{
init_neuro2_workspace(patient_name = NULL, verbose = TRUE)
}
\arguments{
\item{patient_name}{Character string for patient name (optional)}

\item{verbose}{Logical, whether to show verbose output}
}
\value{
TRUE if successful
}
\description{
Sets up directories and loads functionality
}
