% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuro2_llm.R
\name{llm_cache_dir}
\alias{llm_cache_dir}
\title{LLM Cache Directory}
\usage{
llm_cache_dir()
}
\value{
Character string of the cache directory path.
}
\description{
Returns the path to the LLM cache directory, creating it if it doesn't exist.
}
