% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Neuro2MainR6.R
\docType{class}
\name{Neuro2MainR6}
\alias{Neuro2MainR6}
\title{Main Orchestrator for Neuropsychological Reports}
\format{
An R6 class object
}
\description{
Streamlined main orchestrator for neuropsychological report generation.
Provides a simple, user-friendly interface that handles complexity internally.
}
\details{
Neuro2MainR6 Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config}}{Configuration manager object}

\item{\code{error_handler}}{Error handler object}

\item{\code{factory}}{Domain processor factory object}

\item{\code{processors}}{List of domain processors}

\item{\code{status}}{Status tracking list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Neuro2MainR6-new}{\code{Neuro2MainR6$new()}}
\item \href{#method-Neuro2MainR6-setup}{\code{Neuro2MainR6$setup()}}
\item \href{#method-Neuro2MainR6-check_environment}{\code{Neuro2MainR6$check_environment()}}
\item \href{#method-Neuro2MainR6-load_data}{\code{Neuro2MainR6$load_data()}}
\item \href{#method-Neuro2MainR6-process_domains}{\code{Neuro2MainR6$process_domains()}}
\item \href{#method-Neuro2MainR6-process_single_domain}{\code{Neuro2MainR6$process_single_domain()}}
\item \href{#method-Neuro2MainR6-generate_report}{\code{Neuro2MainR6$generate_report()}}
\item \href{#method-Neuro2MainR6-run_full_workflow}{\code{Neuro2MainR6$run_full_workflow()}}
\item \href{#method-Neuro2MainR6-detect_age_group}{\code{Neuro2MainR6$detect_age_group()}}
\item \href{#method-Neuro2MainR6-validate_domains}{\code{Neuro2MainR6$validate_domains()}}
\item \href{#method-Neuro2MainR6-show_summary}{\code{Neuro2MainR6$show_summary()}}
\item \href{#method-Neuro2MainR6-get_available_domains}{\code{Neuro2MainR6$get_available_domains()}}
\item \href{#method-Neuro2MainR6-get_status}{\code{Neuro2MainR6$get_status()}}
\item \href{#method-Neuro2MainR6-clone}{\code{Neuro2MainR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$new(
  config_file = "config.yml",
  variables_file = "_variables.yml",
  verbose = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-setup"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-setup}{}}}
\subsection{Method \code{setup()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$setup()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-check_environment"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-check_environment}{}}}
\subsection{Method \code{check_environment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$check_environment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-load_data"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-load_data}{}}}
\subsection{Method \code{load_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$load_data(
  data_dir = NULL,
  use_duckdb = NULL,
  output_format = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-process_domains"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-process_domains}{}}}
\subsection{Method \code{process_domains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$process_domains(
  domains = NULL,
  age_group = NULL,
  include_multi_rater = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-process_single_domain"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-process_single_domain}{}}}
\subsection{Method \code{process_single_domain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$process_single_domain(processor, domain_key, rater = "self")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-generate_report"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-generate_report}{}}}
\subsection{Method \code{generate_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$generate_report(
  template = NULL,
  output_file = NULL,
  format = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-run_full_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-run_full_workflow}{}}}
\subsection{Method \code{run_full_workflow()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$run_full_workflow(
  domains = NULL,
  age_group = NULL,
  load_data = TRUE,
  generate_report = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-detect_age_group"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-detect_age_group}{}}}
\subsection{Method \code{detect_age_group()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$detect_age_group()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-validate_domains"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-validate_domains}{}}}
\subsection{Method \code{validate_domains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$validate_domains(domains)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-show_summary"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-show_summary}{}}}
\subsection{Method \code{show_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$show_summary()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-get_available_domains"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-get_available_domains}{}}}
\subsection{Method \code{get_available_domains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$get_available_domains()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-get_status"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-get_status}{}}}
\subsection{Method \code{get_status()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$get_status()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Neuro2MainR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Neuro2MainR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Neuro2MainR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
