% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NeuropsychResultsR6.R
\name{NeuropsychResultsR6}
\alias{NeuropsychResultsR6}
\title{NeuropsychResultsR6 Class}
\description{
NeuropsychResultsR6 Class

NeuropsychResultsR6 Class
}
\details{
An R6 class that concatenates and flattens neuropsych results by scale.
This is an R6 implementation of the cat_neuropsych_results function with identical functionality.
}
\section{Methods}{

\describe{
\item{initialize}{Initialize a new NeuropsychResultsR6 object with data and file path.}
\item{process}{Sort the data by percentile, remove duplicates, convert to text, and append to file.}
\item{create_text_placeholder}{Create placeholder text file if it doesn't exist.}
}
}

\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A dataframe containing the data}

\item{\code{file}}{A character string specifying the name of the file}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NeuropsychResultsR6-new}{\code{NeuropsychResultsR6$new()}}
\item \href{#method-NeuropsychResultsR6-create_text_placeholder}{\code{NeuropsychResultsR6$create_text_placeholder()}}
\item \href{#method-NeuropsychResultsR6-process}{\code{NeuropsychResultsR6$process()}}
\item \href{#method-NeuropsychResultsR6-clone}{\code{NeuropsychResultsR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NeuropsychResultsR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-NeuropsychResultsR6-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new NeuropsychResultsR6 object with data and file path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuropsychResultsR6$new(data, file, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A dataframe containing the data}

\item{\code{file}}{A character string specifying the name of the file}

\item{\code{...}}{Additional arguments (ignored)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new NeuropsychResultsR6 object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NeuropsychResultsR6-create_text_placeholder"></a>}}
\if{latex}{\out{\hypertarget{method-NeuropsychResultsR6-create_text_placeholder}{}}}
\subsection{Method \code{create_text_placeholder()}}{
Create a placeholder text file for a given domain if it doesn't already exist.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuropsychResultsR6$create_text_placeholder(domain_file)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{domain_file}}{The domain QMD file (e.g., "_02-01_iq.qmd")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The text filename that was created or already existed
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NeuropsychResultsR6-process"></a>}}
\if{latex}{\out{\hypertarget{method-NeuropsychResultsR6-process}{}}}
\subsection{Method \code{process()}}{
Sort the data by percentile, remove duplicates, convert to text, and write to file as proper QMD content.
Also creates the corresponding text placeholder file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuropsychResultsR6$process()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Invisibly returns NULL after writing to the file.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NeuropsychResultsR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-NeuropsychResultsR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NeuropsychResultsR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
