% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processing_utils.R
\name{process_domain_data}
\alias{process_domain_data}
\title{Process domain data (main processing pipeline)}
\usage{
process_domain_data(pheno, domains)
}
\description{
Process domain data (main processing pipeline)
}
