% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domain_processor_utils_fixed.R
\name{process_multi_rater_domain}
\alias{process_multi_rater_domain}
\title{Process a multi-rater domain}
\usage{
process_multi_rater_domain(
  domain_name,
  data_file,
  age_group = "adult",
  raters = NULL,
  config = NULL
)
}
\arguments{
\item{domain_name}{Domain name}

\item{data_file}{Path to data file}

\item{age_group}{Age group ("adult" or "child")}

\item{raters}{Specific raters to process (NULL for all available)}

\item{config}{Optional configuration}
}
\value{
A list of DomainProcessorR6Combo objects by rater
}
\description{
Process a multi-rater domain
}
