% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm_batch_openai.R
\name{read_prompts_from_dir}
\alias{read_prompts_from_dir}
\title{Read Prompts From Directory of QMD files}
\usage{
read_prompts_from_dir(dir = system.file("prompts", package = "neuro2"))
}
\arguments{
\item{dir}{Directory path (defaults to installed inst/prompts path in the neuro2 package)}
}
\value{
List of prompts with fields name, keyword, text.
}
\description{
Loads prompts from a folder of .qmd files that contain YAML front matter (name, keyword) and a body.
}
