% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_neuro2.R
\name{setup_neuro2}
\alias{setup_neuro2}
\title{Setup neuro2 environment}
\usage{
setup_neuro2(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical, whether to show verbose output}
}
\value{
TRUE if successful
}
\description{
Sets up the neuro2 environment and loads required functionality
}
