% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WorkflowRunnerR6.R
\name{WorkflowRunnerR6}
\alias{WorkflowRunnerR6}
\title{Run the complete neuropsychological workflow}
\description{
Run the complete neuropsychological workflow

Run the complete neuropsychological workflow
}
\examples{

## ------------------------------------------------
## Method `WorkflowRunnerR6$run_workflow`
## ------------------------------------------------

\dontrun{
runner <- WorkflowRunnerR6$new(config)
success <- runner$run_workflow()
runner$print_summary(success)
}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WorkflowRunnerR6-new}{\code{WorkflowRunnerR6$new()}}
\item \href{#method-WorkflowRunnerR6-setup_environment}{\code{WorkflowRunnerR6$setup_environment()}}
\item \href{#method-WorkflowRunnerR6-process_data}{\code{WorkflowRunnerR6$process_data()}}
\item \href{#method-WorkflowRunnerR6-generate_domains}{\code{WorkflowRunnerR6$generate_domains()}}
\item \href{#method-WorkflowRunnerR6-generate_report}{\code{WorkflowRunnerR6$generate_report()}}
\item \href{#method-WorkflowRunnerR6-run_workflow}{\code{WorkflowRunnerR6$run_workflow()}}
\item \href{#method-WorkflowRunnerR6-print_summary}{\code{WorkflowRunnerR6$print_summary()}}
\item \href{#method-WorkflowRunnerR6-clone}{\code{WorkflowRunnerR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-new"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$new(config)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-setup_environment"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-setup_environment}{}}}
\subsection{Method \code{setup_environment()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$setup_environment()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-process_data"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-process_data}{}}}
\subsection{Method \code{process_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$process_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-generate_domains"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-generate_domains}{}}}
\subsection{Method \code{generate_domains()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$generate_domains()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-generate_report"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-generate_report}{}}}
\subsection{Method \code{generate_report()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$generate_report()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-run_workflow"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-run_workflow}{}}}
\subsection{Method \code{run_workflow()}}{
Executes all steps of the workflow in sequence:
\enumerate{
\item Environment setup
\item Data processing
\item Domain file generation
\item Report generation
}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$run_workflow()}\if{html}{\out{</div>}}
}

\subsection{Details}{
Each step is logged and the workflow will stop if any step fails.
The method returns TRUE if all steps complete successfully, FALSE otherwise.
}

\subsection{Returns}{
Logical indicating workflow success (TRUE) or failure (FALSE)
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
runner <- WorkflowRunnerR6$new(config)
success <- runner$run_workflow()
runner$print_summary(success)
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-print_summary"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-print_summary}{}}}
\subsection{Method \code{print_summary()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$print_summary(success)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WorkflowRunnerR6-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WorkflowRunnerR6-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WorkflowRunnerR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
